Type CreateServiceDocRequest as ExtensionObject
